Move needsParens to a language-specific file
